Fix expiration time issues related to Wild Pokemon
Wild pokemon have two fields in the protobuf;
LastModifiedTimestampMs and TimeTillHiddenMs. The
TimeTillHiddenMs will often be a negative number.

Example:
LastModifiedTimestampMs: 1469902449401
TimeTillHiddenMs: -1023634170

To get the actual expire time, you need to add them together.

In addition, you will also see that in a single query you get back
the same Pokemon (same encounter id) in both Wild and Catchable.
The catchable will have a -1 expiration time.

I've modified the CatchablePokemon constructor to calculate the
correct expiration time for WilPokemon and also screen out the
-1 expiration Catchables in MapObjects.addCatchablePokemons()
(They're going to be in Wild array with the correct expiration)