Tutorial state check and enabled (#587)
This will allows new accounts to receive valid responses from PoGo. With the latest changes on PoGo apis, new accounts now require to set the tutorial state to LEGAL_SCREEN.
In addition to this, new accounts require to send the GET_PLAYER requests before mark the tutorial state.
The code check for the presence of the required tutorial state and, if not present, it will set it up.
Feel free to merge or suggest any additional modification required.