Add javadoc redirects for replaced methods
This will allow redirecting to the new classes without the excess
overhead of maintaining the original method content.
The throw will also cause building with the deprecated methods to fail
and redirect the user to the deprecation details.
This is a temporary solution until such time as the methods can be
completely removed in favor of the new structure.